fix: resolve overflow issue in C++ density calculation

+++ cpp_easygraph/classes/operation.cpp
@@ -8,7 +8,7 @@ py::object density(py::object G) {
     if (m == 0 || n <= 1) {
         return py::cast(0);
     }
-    weight_t d = m * 1.0 / (n * (n - 1));
+    weight_t d = static_cast<double>(m) * 1.0 / (static_cast<double>(n) * (n - 1));
     if(G.attr("is_directed")().equal(py::cast(false))){
         d*=2;
     }
